+++ packages/sdk/hardhat.config.ts
@@ -12,7 +12,6 @@ import "./tasks/oracle";
 import "./tasks/plugin";
 import "./tasks/pool";
 import "./tasks/irm";
-import "./tasks/swap/index";
 import "./tasks/fork";
 
 import "./tasks/addChainlinkFeeds";
@@ -31,14 +30,11 @@ import "./tasks/updateFuseFee";
 import "./tasks/upgradePools";
 import "./tasks/replaceDeployer";
 import "./tasks/replacePlugins";
-import "./tasks/one-time/setNonAccruingFlywheels";
 
+import "./tasks/one-time/setNonAccruingFlywheels";
 import "./tasks/one-time/dot-dot-bsc-plugins";
-import  "./tasks/one-time/risedle-pool-upgrade";
-// import "./tasks/one-time/jarvis-polygon-plugins";
 import "./tasks/one-time/jarvis-polygon-mimo-plugin";
 import "./tasks/one-time/arrakis-polygon-plugins";
-import "./tasks/oracle/increase-cardinality";
 
 dotEnvConfig();
