+++ packages/sdk/tasks/oracle/index.ts
@@ -2,3 +2,4 @@ import "./add-tokens";
 import "./get-price";
 import "./set-price";
 import "./update-twap";
+import "./increase-cardinality";
