+++ packages/sdk/tasks/pool/index.ts
@@ -1,1 +1,2 @@
 import "./create";
+import "./non-owner";
